Add Madmate
The Madmate is a member of team Impostors, which has no special abilities.
The Madmate looks like a Crewmate to both other Crewmates and Impostors.
The Impostors do not know who is the Madmate and the Madmate does not know who are the Impostors.

When checking the win condition of the team Impostors, the Madmate is counted as a Crewmate.
The Madmate cannt fix comm and lightout sabotage.